diagnostic_requirements.txt
google-cloud-bigquery>=2.0.0
google-auth>=1.35.0
